WASM Build Exports (No Success)
I want to export the `LZXdecompress` function so it can be used from the JS side of things. What's the most fluent way to get familiar with building for WASM? It seems like all of the tooling around it just doesn't make any sense to me at all. I get the dynamics in which WASM only allows certain features in and out of it's core operation area, I just don't yet understand how you are supposed to bind those to the external JS land, and how you can pass memory between the two. Maybe that's my confusion, you have to build that bridge using WASM's memory storage? That's starting to be what it sounds like. Maybe I can look into how AssemblyScript manages that for you, so I can reverse-implement that in a C-based project.

Learned about Objective Morality today! Or rather, putting that combination of words to the meaning behind it, which I hadn't realized I didn't know how to explain before.

https://www.dynamsoft.com/codepool/emscripten-compile-cc-javascript.html
https://emscripten.org/docs/porting/connecting_cpp_and_javascript/Interacting-with-code.html?highlight=exported_functions#implement-c-in-javascript
https://github.com/toyobayashi/emnapi
https://emscripten.org/docs/porting/connecting_cpp_and_javascript/Interacting-with-code.html?highlight=exported_functions#interacting-with-code-ccall-cwrap
https://www.reddit.com/r/rust/comments/17e4t0p/how_to_return_a_large_arraybuffer_from_a/
https://rustwasm.github.io/wasm-bindgen/api/js_sys/struct.Uint8Array.html#method.view
https://stackoverflow.com/questions/57056652/how-does-webassembly-javascript-memory-interaction-work-with-multiple-typed
https://stackoverflow.com/questions/53602955/using-emscripten-how-to-get-c-uint8-t-array-to-js-blob-or-uint8array/53605865#53605865
https://developer.mozilla.org/en-US/docs/WebAssembly/Loading_and_running
https://nodejs.org/en/learn/getting-started/nodejs-with-webassembly
https://dev.to/sendilkumarn/loading-wasm-as-esm-in-nodejs-2gdb
https://developer.mozilla.org/en-US/docs/WebAssembly/JavaScript_interface/Memory