Decompressor Demo!
I tried running it with a local build, but the file that it outputs is empty
Tried a few attempts at running it with WASM too, but I still have no idea how that works yet

*Which part of the `savegame.dat` does the LZX compression start?*
Managed to get it running, by way of a demo main file that the lzx file had in it

```sh
../../../xnb-unwinder/bin/xnb ./savegame.dat ./savegame-decompressed.dat
```

This was the output for that:

```
3
./savegame.dat
./savegame-decompressed.dat
2
illegal data
```

So yeah I am trying to run this on the demo Xbox 360 savegame file from my Gamedata-Parser project. I think if I trim away part of the header I can successfully decompress the body of the file.